Hiraffe.AdjacencyListRep

Description

A Representation of graphs using adjacency lists.

Documentation

type Graph = GGraph Seq

A graph in which the adjacency lists are stored using Data.Sequence

newtype GGraph f v e

A graph in which the vertices themselves are indices

Constructors

Graph (IntMap (VertexData f v e))

data VertexData f v e

stores the vertex data. the order

Constructors

VertexData !v (IntMap e) (f Int)