Hiraffe.AdjacencyListRep.Map

Description

A Representation of graphs using adjacency lists. The adjacency lists are stored using a Map.

Documentation

type Graph = GGraph Seq

A graph in which the adjacency lists are stored using Data.Sequence

newtype GGraph f i v e

A directed graph, whose vertices store data of type v, and are indexed by type i, and whose edges store data of type e. Edges are indexed by an ordered pair of vertex indices.

The ajdacency lsits are stored using the given f structure.

Constructors

Graph (NEMap i (VertexData f i v e))

data VertexData f i v e

stores the vertex data. the order

Constructors

VertexData !v (Map i e) (f i)