Hiraffe.DFS

Description

Implementation of depth first search.

Documentation

dfs :: (Graph_ graph, Enum (VertexIx graph)) => graph -> VertexIx graph -> TrieF (KV List) (EdgeIx graph) (VertexIx graph)

DFS, from a given vertex.

pre: - the vertex actually occurs in the graph! - the fromEnum instance of a vertex maps the vertices to the range [0,..,V-1]

\(O(V+E)\)

dff :: (Graph_ graph, Enum (VertexIx graph)) => graph -> [TrieF (KV List) (EdgeIx graph) (VertexIx graph)]

Depth first Search (DFS) to discover all components in the graph.

pre: - the fromEnum instance of a vertex maps the vertices to the range [0,..,V-1]

\(O(V+E)\)

dfs'

Arguments

:: (Witherable g, Enum vertexIx)

=> (vertexIx -> g (edge, vertexIx))
Function to generate the out neighbours

-> Int
the number of vertices

-> vertexIx
start vertex

-> TrieF (KV g) edge vertexIx

DFS from the given vertex. The graph is represented by a function that specifies the out vertices.

pre: - the vertex actually occurs in the graph! - the fromEnum instance of a vertex maps the vertices to the range [0,..,V-1]

dff'

Arguments

:: (Witherable f, Witherable g, Enum vertexIx)

=> (vertexIx -> g (edge, vertexIx))
Function to generate the out neighbours

-> Int
the number of vertices

-> f vertexIx
the vertices

-> f (TrieF (KV g) edge vertexIx)

Depth first Search (DFS) to discover all components in the graph.

pre: - the fromEnum instance of a vertex maps the vertices to the range [0,..,V-1]

\(O(V+E)\)