Hiraffe.Graph

Description

Defines various relevant graph classes

Documentation

class HasVertices' graph => HasVertices graph graph' where

Class that expresses that we have a type changing traversal of all vertices.

Methods

vertices :: IndexedTraversal (VertexIx graph) graph graph' (Vertex graph) (Vertex graph')

Traversal of all vertices in the graph

class HasVertices' graph where

A class representing types that have vertices.

Minimal complete definition

vertexAt

Associated Types

type Vertex graph :: Type

Vertices of the graph are of this type

type VertexIx graph :: Type

Vertices are indexed by elements of type VertexIx

Methods

vertexAt :: VertexIx graph -> IndexedTraversal' (VertexIx graph) graph (Vertex graph)

Accessor to a given vertex.

numVertices :: graph -> Int

Number of vertices in the graph.

running time: O(1)

default numVertices :: HasVertices graph graph => graph -> Int

class HasDarts' graph => HasDarts graph graph' where

Class for types that have a type changing traversal of all darts

Methods

darts :: IndexedTraversal (DartIx graph) graph graph' (Dart graph) (Dart graph')

Traversal of all darts in the graph.

class HasDarts' graph where

Class for types that have darts; a dart is a directed edge.

Minimal complete definition

dartAt

Associated Types

type DartIx graph :: Type

Type to index a dart with

type Dart graph :: Type

The data of a dart

Methods

dartAt :: DartIx graph -> IndexedTraversal' (DartIx graph) graph (Dart graph)

Indexed traversal of a given dart.

numDarts :: graph -> Int

Number of darts in the graph.

running time: O(1)

default numDarts :: HasDarts graph graph => graph -> Int

class HasEdges' graph => HasEdges graph graph' where

Class for types that have a type changing traversal of the edges

Methods

edges :: IndexedTraversal (EdgeIx graph) graph graph' (Edge graph) (Edge graph')

Traversal of all edges in the graph

class HasEdges' graph where

A class for things that have edges

Minimal complete definition

edgeAt

Associated Types

type Edge graph :: Type

The edges of the graph are of this type

type EdgeIx graph :: Type

The Edges are indexed by something of type EdgeIx

Methods

edgeAt :: EdgeIx graph -> IndexedTraversal' (EdgeIx graph) graph (Edge graph)

Indexed traversal of a given edge.

numEdges :: graph -> Int

Number of edges in the graph.

running time: O(1)

default numEdges :: HasEdges graph graph => graph -> Int

class HasFaces' graph => HasFaces graph graph' where

Types that have a possibly type changing traversal of all faces.

Methods

faces :: IndexedTraversal (FaceIx graph) graph graph' (Face graph) (Face graph')

Traversal of all faces in the graph

class HasFaces' graph where

A Class for types that have faces.

Minimal complete definition

faceAt

Associated Types

type Face graph

The faces are of type Face graph

type FaceIx graph

Faces are indexed by something of type 'FaceIx graph'

Methods

faceAt :: FaceIx graph -> IndexedTraversal' (FaceIx graph) graph (Face graph)

Indexed traversal of a given face.

numFaces :: graph -> Int

The number of faces in the Planar graph

default numFaces :: HasFaces graph graph => graph -> Int

class (HasVertices graph graph, HasEdges graph graph, DirGraph_ graph) => Graph_ graph where

A graph representing undirected graphs. Note that every undirected graph is also a directed graph.

Methods

fromAdjacencyLists :: (Foldable f, Functor f, Foldable h, Functor h, vi ~ VertexIx graph, v ~ Vertex graph, e ~ Edge graph) => f (vi, v, h (vi, e)) -> graph

Build a graph from its adjacency lists.

If the, in the list of neighbours of vertex u we see a vertex v that itself does not appear in the adjacencylist, we may drop it. In other words if u has a neighbour v, then v better have a specification of its neighbours somewhere.

neighboursOf :: VertexIx graph -> IndexedFold (VertexIx graph) graph (Vertex graph)

All neighbours of a given vertex

incidentEdgesOf :: VertexIx graph -> IndexedFold (EdgeIx graph) graph (Edge graph)

All edges incident to a given vertex

class (HasVertices graph graph, HasDarts graph graph) => DirGraph_ graph where

A class representing directed graphs

Minimal complete definition

endPoints, (outNeighboursOf | outgoingDartsOf)

Methods

endPoints :: graph -> DartIx graph -> (VertexIx graph, VertexIx graph)

Get the endpoints (origin, destination) of an edge

default endPoints :: DartIx graph ~ (VertexIx graph, VertexIx graph) => graph -> DartIx graph -> (VertexIx graph, VertexIx graph)

outNeighboursOf :: VertexIx graph -> IndexedFold (VertexIx graph) graph (Vertex graph)

All outgoing neighbours of a given vertex

outgoingDartsOf :: VertexIx graph -> IndexedFold (DartIx graph) graph (Dart graph)

All outgoing darts incident to a given vertex

default outgoingDartsOf :: DartIx graph ~ (VertexIx graph, VertexIx graph) => VertexIx graph -> IndexedFold (DartIx graph) graph (Dart graph)

class (Graph_ graph, HasFaces graph graph) => PlanarGraph_ graph

A class representing planar graphs