Hiraffe.PlanarGraph

Description

Data type for representing connected planar graphs

Documentation

module Hiraffe.PlanarGraph.Connected

module Hiraffe.PlanarGraph.Class

module Hiraffe.Graph.Class

type PlanarGraph w (s :: k) v e f = PlanarGraphF (CPlanarGraph w) s v e f

A PlanarGraph is essentially a bunch of connected Planargraphs; one for every connected component. These graphs store the global ID's (darts, vertexId's, faceId's) in their data values. This essentially gives us a mapping between the two.

note that a face may actually occur in multiple graphs, hence when we store the edges to the the holes, we store the global edgeId's rather than the local edgeId (dart)'s.

invariant: the outerface has faceId 0

data PlanarGraphF (pg :: Type -> Type -> Type -> Type -> Type) (s :: k) v e f

The Implementation of the PlanarGraph type, where the component is a parameter.

Constructors

PlanarGraph

Fields

_components :: NonEmptyVector (Component pg s)

_rawVertexData :: NonEmptyVector (Raw s (VertexId (Wrap s)) v)

_rawDartData :: NonEmptyVector (Raw s (Dart (Wrap s)) e)

_rawFaceData :: NonEmptyVector (RawFace s f)

type Component (planarGraph :: Type -> Type -> Type -> Type -> Type) (s :: k) = planarGraph (Wrap s) (VertexId s) (Dart s) (FaceId s)

A connected component.

For every face f, and every hole in this face, the facedata points to a dart d on the hole s.t. this dart has the face f on its left. i.e. leftFace d = f

fromConnected' :: forall s v e f. CPlanarGraph Primal s v e f -> Dart s -> PlanarGraph Primal s v e f

Given a (connected) PlanarGraph and a dart that has the outerface on its left | Constructs a planarsubdivision

runningTime: \(O(n)\)