Hiraffe.PlanarGraph.Class

Description

Defines various relevant graph classes

Documentation

class (Graph_ planarGraph, HasFaces planarGraph planarGraph) => PlanarGraph_ planarGraph where

A class representing planar graphs

Minimal complete definition

dualGraph, (incidentFaceOf | leftFaceOf), rightFaceOf, prevDartOf, nextDartOf, boundaryDartOf, boundaryDartOf, boundaryDarts

Associated Types

type DualGraphOf planarGraph

type DualGraphOf planarGraph = planarGraph

type WorldOf planarGraph :: World

type WorldOf planarGraph = Primal

Methods

dualGraph :: planarGraph -> DualGraphOf planarGraph

The dual of this graph

incidentFaceOf :: DartIx planarGraph -> IndexedLens' (FaceIx planarGraph) planarGraph (Face planarGraph)

The face to the left of the dart

leftFaceOf :: DartIx planarGraph -> IndexedLens' (FaceIx planarGraph) planarGraph (Face planarGraph)

The face to the left of the dart. Alternative name for incidentFaceOf

rightFaceOf :: DartIx planarGraph -> IndexedLens' (FaceIx planarGraph) planarGraph (Face planarGraph)

The face to the right of the dart

prevDartOf :: DartIx planarGraph -> IndexedLens' (DartIx planarGraph) planarGraph (Dart planarGraph)

Get the previous edge in order along the face (so ccw for internal faces, and cw for external faces) that is to the left of this dart.

nextDartOf :: DartIx planarGraph -> IndexedLens' (DartIx planarGraph) planarGraph (Dart planarGraph)

Get the next edge in order along the face (so ccw for internal faces, and cw for external faces) that is to the left of this dart.

boundaryDartOf :: FaceIx planarGraph -> IndexedLens' (DartIx planarGraph) planarGraph (Dart planarGraph)

Gets a dart bounding this face. I.e. a dart d such that the face lies to the left of the dart.

boundaryDarts :: FaceIx planarGraph -> planarGraph -> NonEmptyVector (DartIx planarGraph)

The darts bounding this face. The darts are reported in order along the face. This means that for internal faces the darts are reported in *counter clockwise* order along the boundary, whereas for the outer face the darts are reported in clockwise order.

running time: \(O(k)\), where \(k\) is the output size.

boundaryDartsOf :: FaceIx planarGraph -> IndexedFold1 (DartIx planarGraph) planarGraph (Dart planarGraph)

The darts are reported in order along the face. This means that for internal faces the darts are reported in *counter clockwise* order along the boundary, whereas for the outer face the darts are reported in clockwise order.

running time: \(O(k)\), where \(k\) is the output size.

boundaryVertices :: FaceIx planarGraph -> planarGraph -> NonEmptyVector (VertexIx planarGraph)

The vertices bounding this face, for internal faces in counter clockwise order, for the outer face in clockwise order.

running time: \(O(k)\), where \(k\) is the output size.

boundaryVerticesOf :: FaceIx planarGraph -> IndexedFold1 (VertexIx planarGraph) planarGraph (Vertex planarGraph)

The vertices bounding this face, for internal faces in counter clockwise order, for the outer face in clockwise order.

running time: \(O(k)\), where \(k\) is the output size.

class HasFaces' graph => HasFaces graph graph' where

Types that have a possibly type changing traversal of all faces.

Methods

faces :: IndexedTraversal1 (FaceIx graph) graph graph' (Face graph) (Face graph')

Traversal of all faces in the graph

class HasFaces' graph where

A Class for types that have faces.

Minimal complete definition

faceAt

Associated Types

type Face graph

The faces are of type Face graph

type FaceIx graph

Faces are indexed by something of type 'FaceIx graph'

Methods

faceAt :: FaceIx graph -> IndexedTraversal' (FaceIx graph) graph (Face graph)

Indexed traversal of a given face.

numFaces :: graph -> Int

The number of faces in the Planar graph

default numFaces :: HasFaces graph graph => graph -> Int