Hiraffe.PlanarGraph.Component

Description

Connected Components

Documentation

class HasConnectedComponents' graph => HasConnectedComponents graph graph' where

Class that expresses that we have a non-empty type changing traversal of all components.

Methods

connectedComponents :: IndexedTraversal1 (ConnectedComponentIx graph) graph graph' (ConnectedComponent graph) (ConnectedComponent graph')

Traversal of all components in the graph

class HasConnectedComponents' graph where

A class representing types that have components.

Minimal complete definition

connectedComponentAt

Associated Types

type ConnectedComponent graph :: Type

ConnectedComponents of the graph are of this type

type ConnectedComponentIx graph :: Type

ConnectedComponents are indexed by elements of type ConnectedComponentIx

Methods

connectedComponentAt :: ConnectedComponentIx graph -> IndexedTraversal' (ConnectedComponentIx graph) graph (ConnectedComponent graph)

Accessor to a given component.

numConnectedComponents :: graph -> Int

Number of components in the graph.

running time: O(1)

default numConnectedComponents :: HasConnectedComponents graph graph => graph -> Int

data FaceData h f

The Face data consists of the data itself and a list of holes

Constructors

FaceData (Seq h) !f

holes :: Lens (FaceData h f) (FaceData h' f) (Seq h) (Seq h')

lens to access the holes of a face

fData :: Lens (FaceData h f) (FaceData h f') f f'

Lens to access the actual face data

type data Wrap (s :: k)

newtype ComponentId (s :: k)

ComponentId type

Constructors

ComponentId

Fields

unCI :: Int

data Raw s ia a

Helper type for the data that we store in a planar graph

Constructors

Raw !(ComponentId s) !ia !a

dataVal :: Lens (Raw s ia a) (Raw s ia b) a b

get the dataVal of a Raw

data RawFace w (s :: k) f

Face data, if the face is an inner face, store the component and faceId of it. If not, this face must be the outer face (and thus we can find all the face id's it correponds to through the FaceData).

Constructors

RawFace !(Maybe (ComponentId s, FaceIdIn w (Wrap s))) !(FaceData (Dart s) f)

faceIdx :: Lens' (RawFace w s f) (Maybe (ComponentId s, FaceIdIn w (Wrap s)))

Lens to access the faceIx (if it exists)

faceDataVal :: Lens (RawFace w s f) (RawFace w s f') (FaceData (Dart s) f) (FaceData (Dart s) f')

Lens to access the face data