Hiraffe.PlanarGraph.Dart

Description

Data type for representing Darts (edges) in a planar graph.

Documentation

data Dart s

A dart represents a bi-directed edge. I.e. a dart has a direction, however the dart of the oposite direction is always present in the planar graph as well.

Constructors

Dart !(Arc s) !Direction

arc :: Lens' (Dart s) (Arc s)

Arc lens.

direction :: Lens' (Dart s) Direction

Direction lens.

twin :: Dart s -> Dart s

Get the twin of this dart (edge)

>>> twin (dart 0 "+1")
Dart (Arc 0) -1
>>> twin (dart 0 "-1")
Dart (Arc 0) +1

isPositive :: Dart s -> Bool

test if a dart is Positive

asPositive :: Dart s -> Dart s

Returrns the psoitive version of this this ard.

allDarts :: [Dart s]

Enumerates all darts such that allDarts !! i = d = i == fromEnum d

newtype Arc s

An Arc is a directed edge in a planar graph. The type s is used to tie this arc to a particular graph.

Constructors

Arc Int

data Direction

Darts have a direction which is either Positive or Negative (shown as +1 or -1, respectively).

Constructors

Negative

Positive

rev :: Direction -> Direction

Reverse the direcion