Hiraffe.PlanarGraph.EdgeOracle

Description

Data structure to represent a planar graph with which we can test in \(O(1)\) time if an edge between a pair of vertices exists.

Documentation

data EdgeOracle s w e

An Edge Oracle, that lets us efficiently query the data associated with any edge.

main idea: store adjacency lists in such a way that we store an edge (u,v) either in u's adjacency list or in v's. This can be done s.t. all adjacency lists have length at most 6. So every edge, as considered as an undirected edge, would be stored exactly once. (i.e. either at u or at v, but not both)

edgeOracle :: PlanarGraph s w v e f -> EdgeOracle s w (DartId s)

Given a planar graph, construct an edge oracle. Given a pair of vertices this allows us to efficiently find the dart representing this edge in the graph.

pre: No self-loops and no multi-edges!!!

running time: \(O(n)\)

buildEdgeOracle :: forall s w e f g. (Foldable f, Foldable g) => f (VertexIdIn w s, g (VertexIdIn w s, e)) -> EdgeOracle s w e

Builds an edge oracle that can be used to efficiently test if two vertices are connected by an edge.

running time: \(O(n)\)

hasEdge :: VertexIdIn w s -> VertexIdIn w s -> EdgeOracle s w a -> Bool

Test if u and v are connected by an edge.

running time: \(O(1)\)

findEdge :: VertexIdIn w s -> VertexIdIn w s -> EdgeOracle s w a -> Maybe a

Find the edge data corresponding to edge (u,v) if such an edge exists.

running time: \(O(1)\)

findDart :: VertexIdIn w s -> VertexIdIn w s -> EdgeOracle s w (DartId s) -> Maybe (DartId s)

Given a pair of vertices (u,v) returns the dart, oriented from u to v, corresponding to these vertices.

running time: \(O(1)\)

itraverseUndirected :: Applicative f => ((VertexIdIn w s, VertexIdIn w s) -> DartData e -> f (DartData e')) -> EdgeOracle s w e -> f (EdgeOracle s w e')

Traverse the undirected edges stored in the edge oracle

data DartData e

At an (undirected) edge (u,v) we may store either the data of the directed edge (u,v), the data of the directed edge (v,u), or both.

Constructors

UToV e

VToU e

Both

Fields

uToV' :: e

the data associated with the dart in this direction

vToU' :: e

the data associated with the reverse direction