Hiraffe.PlanarGraph.IO

Description

Converting from/to our Adjacency-list representation of the Planar graph

Documentation

toAdjRep :: CPlanarGraph w s v e f -> Gr (Vtx v e) (Face f)

Transforms the planar graph into a format that can be easily converted into JSON format. For every vertex, the adjacent vertices are given in counter-clockwise order.

See toAdjacencyLists for notes on how we handle self-loops.

running time: \(O(n)\)

fromAdjRep :: forall s v e f. Gr (Vtx v e) (Face f) -> CPlanarGraph Primal s v e f

Read a directed planar graph, given in some adjacency list representation into a planar graph. The adjacencylists should be in counter clockwise order.

pre: - the id's are consecutive from 0 to n (where is the number of vertices) - no self-loops and no multi-edges - there is at least one vertex, and at least one edge

running time: \(O(n)\)

buildGraph :: forall s v e. NonEmpty (Vtx v e) -> CPlanarGraph Primal s v e ()

Read a directed planar graph, given by its adjacencylists in counter clockwise order.

pre: - the id's are consecutive from 0 to n (where is the number of vertices) - no self-loops and no multi-edges

running time: \(O(n)\)

fromAdjacencyLists :: forall s w v e g h. (Functor g, Foldable1 g, Foldable1 h, Functor h, Semigroup e) => g (VertexIdIn w s, v, h (VertexIdIn w s, e)) -> CPlanarGraph w s v e ()

Construct an undirected planar graph from a adjacency matrix. For every vertex, all vertices should be given in counter-clockwise order. We will combine the data of negative and positive darts into one using the given semigroup.

pre: No self-loops, and no multi-edges

running time: \(O(n)\).

directedFromAdjacencyLists :: forall s w v e g h. (Functor g, Foldable1 g, Foldable1 h, Functor h) => g (VertexIdIn w s, v, h (VertexIdIn w s, e)) -> CPlanarGraph w s v e ()

Construct a "directed" planar graph from a adjacency matrix. For every vertex, all vertices should be given in counter-clockwise order. The result is directed in the sense that Positive and Negative darts may store different data.

pre: No self-loops, and no multi-edges

running time: \(O(n)\).

reorder :: NonEmptyVector (i, a) -> (i -> Int) -> NonEmptyVector a

make sure we order the data values appropriately

Orphan instances

(FromYAML v, FromYAML e, FromYAML f) => FromYAML (CPlanarGraph 'Primal s v e f)

Instance details

Methods

parseYAML :: Node Pos -> Parser (CPlanarGraph 'Primal s v e f)

(ToYAML v, ToYAML e, ToYAML f) => ToYAML (CPlanarGraph w s v e f)

Instance details

Methods

toYAML :: CPlanarGraph w s v e f -> Node ()

(FromJSON v, FromJSON e, FromJSON f) => FromJSON (CPlanarGraph 'Primal s v e f)

Instance details

Methods

parseJSON :: Value -> Parser (CPlanarGraph 'Primal s v e f)

parseJSONList :: Value -> Parser [CPlanarGraph 'Primal s v e f]

omittedField :: Maybe (CPlanarGraph 'Primal s v e f)

(ToJSON v, ToJSON e, ToJSON f) => ToJSON (CPlanarGraph w s v e f)

Instance details

Methods

toJSON :: CPlanarGraph w s v e f -> Value

toEncoding :: CPlanarGraph w s v e f -> Encoding

toJSONList :: [CPlanarGraph w s v e f] -> Value

toEncodingList :: [CPlanarGraph w s v e f] -> Encoding

omitField :: CPlanarGraph w s v e f -> Bool