Hiraffe.PlanarGraph.Type

Description

Data type for representing possibly disconnected planar graphs

Documentation

data PlanarGraph (w :: World) (s :: k) v e f

A PlanarGraph is essentially a bunch of connected Planargraphs; one for every connected component. These graphs store the global ID's (darts, vertexId's, faceId's) in their data values. This essentially gives us a mapping between the two.

note that a face may actually occur in multiple graphs, hence when we store the edges to the the holes, we store the global edgeId's rather than the local edgeId (dart)'s.

invariants: - the outerface has faceId 0 - the local and global orientation of the darts is consistent

Constructors

PlanarGraph

Fields

_components :: NonEmptyVector (Component w s)

_rawVertexData :: NonEmptyVector (Raw s (VertexIx (Component w s)) v)

_rawDartData :: NonEmptyVector (Raw s (DartIx (Component w s)) e)

_rawFaceData :: NonEmptyVector (RawFace w s f)

type Component w (s :: k) = CPlanarGraph w (Wrap s) (VertexId s) (Dart s) (FaceId s)

A connected component.

For every face f, and every hole in this face, the facedata points to a dart d on the hole s.t. this dart has the face f on its left. i.e. leftFace d = f

components :: Lens' (PlanarGraph w s v e f) (NonEmptyVector (Component w s))

Lens to access the connected components of a planar subdivision.

rawDartData :: Lens (PlanarGraph w s v e f) (PlanarGraph w s v e' f) (NonEmptyVector (Raw s (DartIx (Component w s)) e)) (NonEmptyVector (Raw s (DartIx (Component w s)) e'))

Lens to access the raw dart daat a

rawVertexData :: Lens (PlanarGraph w s v e f) (PlanarGraph w s v' e f) (NonEmptyVector (Raw s (VertexIx (Component w s)) v)) (NonEmptyVector (Raw s (VertexIx (Component w s)) v'))

Lens to access the raw vertex data

rawFaceData :: Lens (PlanarGraph w s v e f) (PlanarGraph w s v e f') (NonEmptyVector (RawFace w s f)) (NonEmptyVector (RawFace w s f'))

Access the raw face data

component :: ComponentId s -> Lens' (PlanarGraph w s v e f) (Component w s)

Lens to access a particular component of the planar subdivision.

fromConnected' :: forall w s v e f. CPlanarGraph w s v e f -> Dart s -> PlanarGraph w s v e f

Given a (connected) PlanarGraph and a dart that has the outerface on its left | Constructs a PlanarGraph

runningTime: \(O(n)\)

asLocalD :: DartIx (PlanarGraph w s v e f) -> PlanarGraph w s v e f -> (ComponentId s, DartIx (Component w s), Component w s)

Given a "global" dart id, get the component and local dart info

asLocalV :: VertexIx (PlanarGraph w s v e f) -> PlanarGraph w s v e f -> (ComponentId s, VertexIx (Component w s), Component w s)

Given a global vertexId, get the local info

itraverseEdges1 :: forall g s a e e'. Apply g => (Dart s -> e -> g e') -> NonEmptyVector (Raw s a e) -> g (NonEmptyVector (Raw s a e'))

itraverse the edges; i.e. makes sure to only apply our function to the positive darts.