Hiraffe.ShortestPath.Dijkstra

Description

Implementation of Dijkstra's shortest path algorithm.

Documentation

shortestPaths

Arguments

:: forall graph adjs v w. (Foldable graph, Foldable adjs, Ord v, Ord w, Monoid w)

=> (v -> v -> Top w)
weight function

-> v
source

-> graph (v, adjs v)
adjacency graph

-> [(v, Top w)]

Dijkstra's algorithm for computing the shortest paths from the given source node to all other nodes in the unweighted graph

O(Vlog V + Elog V)

data WithPath f v r

Constructors

WithPath

Fields

path :: f v

dist :: !r