Hiraffe.ShortestPath.FloydWarshall

Description

Implementation of Floyd-Warshall shortest path algorithm.

See Wikipedia article for details: https://en.wikipedia.org/wiki/Floyd%E2%80%93Warshall_algorithm

Documentation

allPairsShortestPaths

Arguments

:: (HasVertices' graph, VertexIx graph ~ Int, Num dist, Ord dist)

=> (VertexIx graph -> VertexIx graph -> Top dist)
edge/dart lengths

-> graph

-> Array (VertexIx graph, VertexIx graph) (Top dist)

All pair shortest paths. See allPiarsShortestPathsWith for more info.

running time: \(O(n^3)\), where \(n\) is the number of vertices of the graph.

allPairsShortestPathsWithNext

Arguments

:: (HasVertices' graph, VertexIx graph ~ Int, Num dist, Ord dist)

=> (VertexIx graph -> VertexIx graph -> Top dist)
edge/dart lengths

-> graph

-> Array (VertexIx graph, VertexIx graph) (Top (WithNext (VertexIx graph) dist))

All pair shortest paths. Annotates every vertex with its successor vertex along the path. See allPiarsShortestPathsWith for more info.

running time: \(O(n^3)\), where \(n\) is the number of vertices of the graph.

allPairsShortestPathsWith

Arguments

:: (HasVertices' graph, VertexIx graph ~ Int, Semigroup dist, Ord dist)

=> (VertexIx graph -> VertexIx graph -> Top dist)
edge/dart lengths

-> graph

-> Array (VertexIx graph, VertexIx graph) (Top dist)

All pair shortest paths.

Note that essentially all information in the graph is actually encoded in the given distance function; i.e. the length function may be called for any pair of vertices; even if they are not connected by an edge.

running time: \(O(n^3)\), where \(n\) is the number of vertices of the graph.

extractShortestPath

Arguments

:: (Eq vertexIx, Ix vertexIx)

=> Array (vertexIx, vertexIx) (Top (WithNext vertexIx dist))

-> vertexIx
source vertex

-> vertexIx
target vertex

-> Top (dist, NonEmpty vertexIx)

Extracts a shortest path from the table.

data WithNext vertex dist

Keep track of the next vertex along the path.

Constructors

WithNext !vertex !dist