Hiraffe.Graph

Description

Defines various relevant graph classes

Documentation

class HasVertices' graph => HasVertices graph graph' where

Class that expresses that we have a non-empty type changing traversal of all vertices.

Methods

vertices :: IndexedTraversal1 (VertexIx graph) graph graph' (Vertex graph) (Vertex graph')

Traversal of all vertices in the graph

class HasVertices' graph where

A class representing types that have vertices.

Minimal complete definition

vertexAt

Associated Types

type Vertex graph :: Type

Vertices of the graph are of this type

type VertexIx graph :: Type

Vertices are indexed by elements of type VertexIx

Methods

vertexAt :: VertexIx graph -> IndexedTraversal' (VertexIx graph) graph (Vertex graph)

Accessor to a given vertex.

numVertices :: graph -> Int

Number of vertices in the graph.

running time: O(1)

default numVertices :: HasVertices graph graph => graph -> Int

class HasDarts' graph => HasDarts graph graph' where

Class for types that have a type changing traversal of all darts

Methods

darts :: IndexedTraversal (DartIx graph) graph graph' (Dart graph) (Dart graph')

Traversal of all darts in the graph.

class HasDarts' graph where

Class for types that have darts; a dart is a directed edge.

Minimal complete definition

dartAt

Associated Types

type DartIx graph :: Type

Type to index a dart with

type Dart graph :: Type

The data of a dart

Methods

dartAt :: DartIx graph -> IndexedTraversal' (DartIx graph) graph (Dart graph)

Indexed traversal of a given dart.

numDarts :: graph -> Int

Number of darts in the graph.

running time: O(1)

default numDarts :: HasDarts graph graph => graph -> Int

class HasEdges' graph => HasEdges graph graph' where

Class for types that have a type changing traversal of the edges

Methods

edges :: IndexedTraversal (EdgeIx graph) graph graph' (Edge graph) (Edge graph')

Traversal of all edges in the graph

class HasEdges' graph where

A class for things that have edges

Minimal complete definition

edgeAt

Associated Types

type Edge graph :: Type

The edges of the graph are of this type

type EdgeIx graph :: Type

The Edges are indexed by something of type EdgeIx

Methods

edgeAt :: EdgeIx graph -> IndexedTraversal' (EdgeIx graph) graph (Edge graph)

Indexed traversal of a given edge.

numEdges :: graph -> Int

Number of edges in the graph.

running time: O(1)

default numEdges :: HasEdges graph graph => graph -> Int

class (HasVertices graph graph, HasEdges graph graph, BidirGraph_ graph) => Graph_ graph where

A class representing non-empty undirected graphs. Note that every undirected graph is also a directed graph.

Associated Types

type GraphFromAdjListExtraConstraints graph (h :: Type -> Type) :: Constraint

Possible additional constraints for constructing a DirGraph

type GraphFromAdjListExtraConstraints graph h = ()

Methods

fromAdjacencyLists :: (Foldable1 f, Functor f, Foldable h, Functor h, vi ~ VertexIx graph, v ~ Vertex graph, e ~ Edge graph, GraphFromAdjListExtraConstraints graph h) => f (vi, v, h (vi, e)) -> graph

Build a graph from its adjacency lists.

If the, in the list of neighbours of vertex u we see a vertex v that itself does not appear in the adjacencylist, we may drop it. In other words if u has a neighbour v, then v better have a specification of its neighbours somewhere.

neighboursOf :: VertexIx graph -> IndexedFold (VertexIx graph) graph (Vertex graph)

All neighbours of a given vertex

incidentEdgesOf :: VertexIx graph -> IndexedFold (EdgeIx graph) graph (Edge graph)

All edges incident to a given vertex

class (HasVertices graph graph, HasDarts graph graph) => DiGraph_ graph where

A class representing non-empty directed graphs

Minimal complete definition

diGraphFromAdjacencyLists, (endPoints | headOf, tailOf), (outNeighboursOf | outgoingDartsOf), twinDartOf

Associated Types

type DiGraphFromAdjListExtraConstraints graph (h :: Type -> Type) :: Constraint

Possible additional constraints for constructing a DirGraph

type DiGraphFromAdjListExtraConstraints graph h = ()

Methods

diGraphFromAdjacencyLists :: (Foldable1 f, Functor f, Foldable h, Functor h, vi ~ VertexIx graph, v ~ Vertex graph, d ~ Dart graph, DiGraphFromAdjListExtraConstraints graph h) => f (vi, v, h (vi, d)) -> graph

Build a directed graph from its adjacency lists.

endPoints :: graph -> DartIx graph -> (VertexIx graph, VertexIx graph)

Get the endpoints (origin, destination) of a dart

endPointsOf :: DartIx graph -> IndexedGetter (VertexIx graph, VertexIx graph) graph (Vertex graph, Vertex graph)

Given a dart, produce an indexed getter to access the endpoints (u,v) of the dart. where u is the origintail, and v is the destinationhead.

outNeighboursOf :: VertexIx graph -> IndexedFold (VertexIx graph) graph (Vertex graph)

All outgoing neighbours of a given vertex

outgoingDartsOf :: VertexIx graph -> IndexedFold (DartIx graph) graph (Dart graph)

All outgoing darts incident to a given vertex

default outgoingDartsOf :: DartIx graph ~ (VertexIx graph, VertexIx graph) => VertexIx graph -> IndexedFold (DartIx graph) graph (Dart graph)

twinDartOf :: DartIx graph -> Getter graph (Maybe (DartIx graph))

The twin of this dart, if it exits

headOf :: DartIx graph -> IndexedGetter (VertexIx graph) graph (Vertex graph)

The vertex this dart is heading in to (i.e. its destination.)

tailOf :: DartIx graph -> IndexedGetter (VertexIx graph) graph (Vertex graph)

The tail of a dart, i.e. the vertex this dart is leaving from (i.e. its origin)

class DiGraph_ graph => BidirGraph_ graph where

Types representing non-empty bidirected graphs, i.e. a directed graph, but all directed edges are guaranteed to exist in both directions.

Methods

twinOf :: DartIx graph -> Getter graph (DartIx graph)

The twin of this dart.

getPositiveDart :: graph -> EdgeIx graph -> DartIx graph

Given An edgeIx, gets the positive dart that (together with its twin) represents this edge.