Hiraffe.PlanarGraph.Class

Description

Defines various relevant graph classes

Documentation

class (Graph_ planarGraph, HasFaces planarGraph planarGraph) => PlanarGraph_ planarGraph where

A class representing planar graphs

Minimal complete definition

dualGraph, (incidentFaceOf | leftFaceOf), _DualFaceIx, _DualVertexIx, prevDartOf, nextDartOf, boundaryDartOf, boundaryDartsFrom

Associated Types

type DualGraphOf planarGraph

type DualGraphOf planarGraph = planarGraph

type WorldOf planarGraph :: World

type WorldOf planarGraph = Primal

Methods

dualGraph :: planarGraph -> DualGraphOf planarGraph

The dual of this graph

_DualFaceIx :: proxy planarGraph -> Iso' (VertexIx planarGraph) (FaceIx (DualGraphOf planarGraph))

Convert from a VertexIx in the primal graph to a FaceIx in the dual graph.

(Likely this will just be the identiy function or a simple coercion)

_DualVertexIx :: proxy planarGraph -> Iso' (FaceIx planarGraph) (VertexIx (DualGraphOf planarGraph))

Convert from a FaceIx in the primal graph to a VertexIx in the dual graph.

(Likely this will just be the identiy function or a simple coercion)

incidentFaceOf :: DartIx planarGraph -> IndexedLens' (FaceIx planarGraph) planarGraph (Face planarGraph)

The face to the left of the dart

leftFaceOf :: DartIx planarGraph -> IndexedLens' (FaceIx planarGraph) planarGraph (Face planarGraph)

The face to the left of the dart. Alternative name for incidentFaceOf

rightFaceOf :: DartIx planarGraph -> IndexedLens' (FaceIx planarGraph) planarGraph (Face planarGraph)

The face to the right of the dart

prevDartOf :: DartIx planarGraph -> IndexedLens' (DartIx planarGraph) planarGraph (Dart planarGraph)

Get the previous edge in order along the face (so ccw for internal faces, and cw for external faces) that is to the left of this dart.

nextDartOf :: DartIx planarGraph -> IndexedLens' (DartIx planarGraph) planarGraph (Dart planarGraph)

Get the next edge in order along the face (so ccw for internal faces, and cw for external faces) that is to the left of this dart.

boundaryDartOf :: FaceIx planarGraph -> IndexedLens' (DartIx planarGraph) planarGraph (Dart planarGraph)

Gets a dart bounding this face. I.e. a dart d such that the face lies to the left of the dart.

boundaryDartsFrom :: DartIx planarGraph -> IndexedFold1 (DartIx planarGraph) planarGraph (Dart planarGraph)

Given some dart d, compute the darts along the face containing d (starting with d). I.e. we iterate nextDartOf until we are back at the starting dart.

Note that - if the dart is on the outerBoundary of a face, this reports the darts in CCW order - if the dart is on an an inner component, this means we report the darts in CW order.

class HasFaces' graph => HasFaces graph graph' where

Types that have a possibly type changing traversal of all faces.

Methods

faces :: IndexedTraversal1 (FaceIx graph) graph graph' (Face graph) (Face graph')

Traversal of all faces in the graph

class HasFaces' graph where

A Class for types that have faces.

Minimal complete definition

faceAt

Associated Types

type Face graph

The faces are of type Face graph

type FaceIx graph

Faces are indexed by something of type 'FaceIx graph'

Methods

faceAt :: FaceIx graph -> IndexedTraversal' (FaceIx graph) graph (Face graph)

Indexed traversal of a given face.

numFaces :: graph -> Int

The number of faces in the Planar graph

default numFaces :: HasFaces graph graph => graph -> Int

class (DiGraph_ planarGraph, HasFaces planarGraph planarGraph) => HasOuterBoundaryOf planarGraph where

A class for planar graphs for which we can report the outer boundary of a face.

Minimal complete definition

outerBoundaryDarts

Methods

outerBoundaryDarts :: FaceIx planarGraph -> planarGraph -> NonEmptyVector (DartIx planarGraph)

The darts on the outer boundary of this face, in counter clockwise order along the boundary.

pre: faceIx /= outerFaceIx

running time: \(O(k)\), where \(k\) is the output size.

outerBoundaryDartsOf :: FaceIx planarGraph -> IndexedFold1 (DartIx planarGraph) planarGraph (Dart planarGraph)

The darts on the outer boundary of the face, in counter clockwise order along the boundary.

pre: faceIx /= outerFaceIx

running time: \(O(k)\), where \(k\) is the output size.

outerBoundaryVertices :: FaceIx planarGraph -> planarGraph -> NonEmptyVector (VertexIx planarGraph)

The vertices on the outer boundary of the face, in counter clockwise order along the boundary.

pre: faceIx /= outerFaceIx

running time: \(O(k)\), where \(k\) is the output size.

outerBoundaryVerticesOf :: FaceIx planarGraph -> IndexedFold1 (VertexIx planarGraph) planarGraph (Vertex planarGraph)

The vertices on the outer boundary of the face, in counter clockwise order along the boundary.

running time: \(O(k)\), where \(k\) is the output size.

class PlanarGraph_ planeGraph => HasOuterFace planeGraph where

A class for graphs that have an outer face (and conversely, that have inner faces).

Minimal complete definition

outerFaceId | outerFaceDart

Methods

outerFace :: Eq (FaceIx planeGraph) => IndexedLens' (FaceIx planeGraph) planeGraph (Face planeGraph)

Getter to access the outer face

interiorFaces :: Eq (FaceIx planeGraph) => IndexedTraversal' (FaceIx planeGraph) planeGraph (Face planeGraph)

Traversal of all interior faces in the graph

outerFaceId :: planeGraph -> FaceIx planeGraph

gets the id of the outer face

outerFaceDart :: planeGraph -> DartIx planeGraph

gets a dart incident to the outer face (in particular, that has the outerface on its left)

class HasInnerComponent planarGraph where

Class stating that some type can report the inner components of a face.

Methods

innerComponentsAt :: FaceIx planarGraph -> IndexedFold (DartIx planarGraph) planarGraph (Dart planarGraph)

Given a faceIx fi get a Fold of the inner components of the face; i.e. we are given one dart for every inner component. the dart has the face with index fi to its left.