Hiraffe.PlanarGraph.Class

Description

Defines various relevant graph classes

Documentation

class (Graph_ planarGraph, HasFaces planarGraph planarGraph) => PlanarGraph_ planarGraph where

A class representing planar graphs

Minimal complete definition

dualGraph, leftFace, rightFace, prevEdge, nextEdge, boundaryDart, boundary

Associated Types

type DualGraphOf planarGraph

type DualGraphOf planarGraph = planarGraph

Methods

dualGraph :: planarGraph -> DualGraphOf planarGraph

The dual of this graph

leftFace :: DartIx planarGraph -> planarGraph -> FaceIx planarGraph

The face to the left of the dart

rightFace :: DartIx planarGraph -> planarGraph -> FaceIx planarGraph

The face to the right of the dart

prevEdge :: DartIx planarGraph -> planarGraph -> DartIx planarGraph

Get the previous edge in order along the face (so ccw for internal faces, and cw for external faces ) that is to the left of this dart.

nextEdge :: DartIx planarGraph -> planarGraph -> DartIx planarGraph

Get the next edge in order along the face (so ccw for internal faces, and cw for external faces ) that is to the left of this dart.

boundaryDart :: FaceIx planarGraph -> planarGraph -> DartIx planarGraph

Gets a dart bounding this face. I.e. a dart d such that the face lies to the left of the dart.

boundary :: FaceIx planarGraph -> planarGraph -> Vector (DartIx planarGraph)

The darts bounding this face. The darts are reported in order along the face. This means that for internal faces the darts are reported in *counter clockwise* order along the boundary, whereas for the outer face the darts are reported in clockwise order.

boundaryVertices :: FaceIx planarGraph -> planarGraph -> Vector (VertexIx planarGraph)

The vertices bounding this face, for internal faces in counter clockwise order, for the outer face in clockwise order.

class HasFaces' graph => HasFaces graph graph' where

Types that have a possibly type changing traversal of all faces.

Methods

faces :: IndexedTraversal (FaceIx graph) graph graph' (Face graph) (Face graph')

Traversal of all faces in the graph

class HasFaces' graph where

A Class for types that have faces.

Minimal complete definition

faceAt

Associated Types

type Face graph

The faces are of type Face graph

type FaceIx graph

Faces are indexed by something of type 'FaceIx graph'

Methods

faceAt :: FaceIx graph -> IndexedTraversal' (FaceIx graph) graph (Face graph)

Indexed traversal of a given face.

numFaces :: graph -> Int

The number of faces in the Planar graph

default numFaces :: HasFaces graph graph => graph -> Int